services/assistant/assistant_api/pyproject.toml
[tool.poetry]
name = "Assistant"
version = "0.1.0"
description = ""
authors = ["Maksim Kosinov <mkosinov@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
pydantic-settings = "^2.3.4"
gunicorn = "^22.0.0"
beanie = "^1.26.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.5.0"
pytest = "^8.3.1"

[tool.isort]
src_paths = ["src", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
